Moved data classes to their own file
