+++ Example/Tests/Tests.swift
@@ -8,7 +8,7 @@ class MetricSpec: QuickSpec {
     override func spec() {
         describe("test metric creation") {
             context("can send different kinds of metrics") {
-                let m: Metric.MetricData = Metric.MetricData(host: nil, tags: [], metric_name:"test.metric", type: Metric.MetricData.MetricType.gauge, points: [DataPoint(timestamp: TimeInterval(1525377826.2537289), value: 1001.1)])
+                let m = MetricData(host: nil, tags: [], metric_name:"test.metric", type: MetricData.MetricType.gauge, points: [DataPoint(timestamp: TimeInterval(1525377826.2537289), value: 1001.1)])
                 it("can create a basic metric") {
                     expect(m.metric_name).to(equal("test.metric"))
                 }
@@ -24,9 +24,9 @@ class MetricSpec: QuickSpec {
                 }
             }
             context ("can send multiple metrics") {
-                let gauge_metric: Metric.MetricData = Metric.MetricData(host: nil, tags: ["test:1"], metric_name:"test.metric1", type: Metric.MetricData.MetricType.gauge, points: [DataPoint(timestamp: TimeInterval(1525377826.2537289), value: 1)])
-                let rate_metric: Metric.MetricData = Metric.MetricData(host: "device:fun_ios", tags: ["test:2"], metric_name:"test.metric2", type: Metric.MetricData.MetricType.rate(10), points: [DataPoint(timestamp: TimeInterval(1525377828.2537289), value: 2)])
-                let count_metric: Metric.MetricData = Metric.MetricData(host: "device:another_device", tags: ["test:3"], metric_name:"test.metric3", type: Metric.MetricData.MetricType.count(100), points: [DataPoint(timestamp: TimeInterval(1525377820.2537289), value: 3)])
+                let gauge_metric = MetricData(host: nil, tags: ["test:1"], metric_name:"test.metric1", type: MetricData.MetricType.gauge, points: [DataPoint(timestamp: TimeInterval(1525377826.2537289), value: 1)])
+                let rate_metric = MetricData(host: "device:fun_ios", tags: ["test:2"], metric_name:"test.metric2", type: MetricData.MetricType.rate(10), points: [DataPoint(timestamp: TimeInterval(1525377828.2537289), value: 2)])
+                let count_metric = MetricData(host: "device:another_device", tags: ["test:3"], metric_name:"test.metric3", type: MetricData.MetricType.count(100), points: [DataPoint(timestamp: TimeInterval(1525377820.2537289), value: 3)])
                 Datadog.metric.send(series: [gauge_metric, rate_metric, count_metric])
                 it("can serialize to json") {
                     let encoder = JSONEncoder()
@@ -47,7 +47,7 @@ class EventSpec: QuickSpec {
     override func spec() {
         describe("test event creation") {
             context("can send different kinds of metrics") {
-                let e: Event.EventData = Event.EventData(host: "ios", tags:[], title: "test title", text: "test text", date_happened: 1525412871, priority: .normal, alert_type: .info, aggregation_key: nil, source_type_name: nil)
+                let e: EventData = EventData(host: "ios", tags:[], title: "test title", text: "test text", date_happened: 1525412871, priority: .normal, alert_type: .info, aggregation_key: nil, source_type_name: nil)
                 it("can create a basic metric") {
                     expect(e.title).to(equal("test title"))
                 }
